+++ directio-project/asakusa-directio-dmdl/src/main/java/com/asakusafw/dmdl/directio/csv/driver/CsvFieldDriver.java
@@ -45,12 +45,12 @@ public class CsvFieldDriver  extends PropertyAttributeDriver {
     /**
      * The attribute name.
      */
-    public static final String TARGET_NAME = "directio.csv.field";
+    public static final String TARGET_NAME = "directio.csv.field"; //$NON-NLS-1$
 
     /**
      * The element name.
      */
-    public static final String ELEMENT_NAME = "name";
+    public static final String ELEMENT_NAME = "name"; //$NON-NLS-1$
 
     @Override
     public String getTargetName() {
@@ -80,7 +80,7 @@ public class CsvFieldDriver  extends PropertyAttributeDriver {
         environment.report(new Diagnostic(
                 Level.ERROR,
                 attribute,
-                "Other @directio.csv.* is already declared in {0}.{1}",
+                Messages.getString("CsvFieldDriver.diagnosticDuplicateModelAttribute"), //$NON-NLS-1$
                 declaration.getOwner().getName().identifier,
                 declaration.getName().identifier));
         return false;
@@ -108,7 +108,7 @@ public class CsvFieldDriver  extends PropertyAttributeDriver {
         environment.report(new Diagnostic(
                 Level.ERROR,
                 attribute,
-                "Invalid type for @{2} ({0}.{1}), must be one of {3}",
+                Messages.getString("CsvFieldDriver.diagnosticInvalidType"), //$NON-NLS-1$
                 declaration.getOwner().getName().identifier,
                 declaration.getName().identifier,
                 attribute.name.toString(),